월간 리포트 분석 요청 기능 추가

+++ src/main/java/com/nutriassistant/nutriassistant_back/DTO/MealDto.java
@@ -9,8 +9,8 @@ public record MealDto(
         String main2,
         String side,
         String kimchi,
-        Integer kcal,
-        Integer prot
+        Double kcal,
+        Double prot
 ) {
     public static MealDto from(MealPlanMenu m) {
         if (m == null) return null;
